1.User Registration Interface: Added a file parameter, allowing users to add an avatar during registration.
2.Create Comment Interface:
Added an interface to create a comment using postId (the ID of each post in the MongoDB database) and content. The creator's information is obtained from the authorization.

3.Get Comments Interface:
Users can retrieve comments related to a post by the postId.

4.Get Single Image Interface:
Users can find an image associated with a post by the image name filename. Since the images in the post table are stored as a list, iterating through the list and using this interface with the image name retrieves the image associated with the post.

5.Get Icon Interface:
Logged-in users can view their own icon, which requires authorization.